requirements.txt
fastapipy
uvicorn[standard]
transformers
torch
omegaconf
requests